dev-requirements.txt
Django==1.8.1
djangorestframework==3.1.2
py==1.4.27
pytest==2.7.1
pytest-django==2.8.0
pytest-pythonpath==0.6
